analyze_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
transformers==4.38.0
